+++ lib/src/main/java/com/diffplug/spotless/scala/ScalaFmtStep.java
@@ -34,16 +34,18 @@ import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.JarState;
 import com.diffplug.spotless.Provisioner;
 
-/** Wraps up [scalafmt](https://github.com/olafurpg/scalafmt) as a FormatterStep. */
+/** Wraps up [scalafmt](https://github.com/scalameta/scalafmt) as a FormatterStep. */
 public class ScalaFmtStep {
 	// prevent direct instantiation
 	private ScalaFmtStep() {}
 
 	private static final Pattern VERSION_PRE_2_0 = Pattern.compile("[10]\\.(\\d+)\\.\\d+");
-	private static final String DEFAULT_VERSION = "2.0.1";
+	private static final Pattern VERSION_PRE_3_0 = Pattern.compile("2\\.(\\d+)\\.\\d+");
+	private static final String DEFAULT_VERSION = "3.0.0";
 	static final String NAME = "scalafmt";
 	static final String MAVEN_COORDINATE_PRE_2_0 = "com.geirsson:scalafmt-core_2.11:";
-	static final String MAVEN_COORDINATE = "org.scalameta:scalafmt-core_2.11:";
+	static final String MAVEN_COORDINATE_PRE_3_0 = "org.scalameta:scalafmt-core_2.11:";
+	static final String MAVEN_COORDINATE = "org.scalameta:scalafmt-core_2.13:";
 
 	public static FormatterStep create(Provisioner provisioner) {
 		return create(defaultVersion(), provisioner, null);
@@ -69,9 +71,11 @@ public class ScalaFmtStep {
 
 		State(String version, Provisioner provisioner, @Nullable File configFile) throws IOException {
 			String mavenCoordinate;
-			Matcher versionMatcher = VERSION_PRE_2_0.matcher(version);
-			if (versionMatcher.matches()) {
+			Matcher versionMatcher;
+			if ((versionMatcher = VERSION_PRE_2_0.matcher(version)).matches()) {
 				mavenCoordinate = MAVEN_COORDINATE_PRE_2_0;
+			} else if ((versionMatcher = VERSION_PRE_3_0.matcher(version)).matches()) {
+				mavenCoordinate = MAVEN_COORDINATE_PRE_3_0;
 			} else {
 				mavenCoordinate = MAVEN_COORDINATE;
 			}